Fix Option<Decimal> serde round-trip through serde_json::Value
Extract shared decimal_float_option module into tensorzero-types
and use it in both cost.rs and usage.rs. This works around the
upstream rust_decimal issue where #[serde(flatten)] causes null
to go through visit_unit() instead of visit_none().
